Migrate cafe session data to cafes redux
