+++ App/features/cafes/index.ts
@@ -1,5 +1,13 @@
 import * as cafesActions from './actions'
 import cafesReducer, { CafesState, CafesAction } from './reducer'
+import * as cafesSelectors from './selectors'
 import cafesSaga from './sagas'
 
-export { cafesActions, cafesReducer, CafesState, CafesAction, cafesSaga }
+export {
+  cafesActions,
+  cafesReducer,
+  cafesSelectors,
+  CafesState,
+  CafesAction,
+  cafesSaga
+}
